diploma/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1
django-allauth==0.51.0
djangorestframework==3.13.1
drf-spectacular==0.24.2
drf-spectacular-sidecar==2022.10.1
idna==3.3
inflection==0.5.1
jsonschema==4.16.0
kombu==5.2.4
oauthlib==3.2.2
packaging==21.3
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1